Connection timeout

+++ manhole.sh
@@ -13,14 +13,14 @@ EOF
 fi
 source ./manhole-ordering.txt
 
-curl --fail -s -o ./manhole-script -D ./manhole-headers "${manholeurl}/${stationid}"
+curl --max-time 20 --fail -s -o ./manhole-script -D ./manhole-headers "${manholeurl}/${stationid}"
 status="$?"
 ordering="$(grep Ordering ./manhole-headers | sed -e "s+.*: *++g" | tr -d "\r")"
 
 if [ -e "$statusled" ]; then
-  if [ "$status" == "0" ] || [ "$status" == "22" ]; then
-    echo -n "$((($(cat "$statusled") + 1) % 2))" > "$statusled"
-  fi
+    if [ "$status" == "0" ] || [ "$status" == "22" ]; then
+        echo -n "$((($(cat "$statusled") + 1) % 2))" > "$statusled"
+    fi
 fi
 
 if [ "$status" == "0" ] && ((ordering > last_ordering)); then
@@ -28,9 +28,11 @@ if [ "$status" == "0" ] && ((ordering > last_ordering)); then
     now="$(date -Iseconds)"
     chmod ugo+x ./manhole-script
     ./manhole-script > ./manhole-log.txt 2>&1
-    curl --fail -s -X PUT --data-binary @./manhole-log.txt "${manholeurl}/${stationid}/${ordering}"
-
-    cat > ./manhole-ordering.txt <<EOF
+    if curl --max-time 20 --fail -s -X PUT --data-binary @./manhole-log.txt "${manholeurl}/${stationid}/${ordering}"; then
+        cat > ./manhole-ordering.txt <<EOF
 last_ordering="$ordering"
 EOF
+    else
+        echo "Sending results failed. Script will run again..."
+    fi
 fi
